+++ changelog.md
@@ -4,7 +4,11 @@
 
 ### Added
 
+- Added 2 new places for the key
+-> an environment variable from .env-file
+-> a server variable in the hosting
 - Added the function debug() to return the actual configuration for debugging purposes
+- Added some documentation
 
 ### Changed